XM88 Blue Cursor (#9739)
# About the pull request

Standardizes crosshair color - converts the red XM88 crosshair to a blue
one like all other custom crosshairs.

The sprite is the same as the old one in #2291, just recolored to match
the pattern of the current blue ones.


![image](https://github.com/user-attachments/assets/ddca5f7a-a275-4eaf-8c31-491cab50fe13)

# Explain why it's good for the game

Visibility is good. To compare the current **RED** crosshair color to
the average resin color:

![image](https://github.com/user-attachments/assets/bc581032-e829-456f-9146-230b4ed8f95c)

And to compare it to the **BLUE** ones:

![image](https://github.com/user-attachments/assets/d1b0927d-ffc2-4250-9dea-cc05fe97b358)

Now this is technically for front-end accessibility, I tend to also like
following it for UI items.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/4b4289b0-dc29-4fad-82a5-5953d574c99d

</details>



# Changelog
:cl: matttheficus, harryob
ui: XM88 custom crosshair is now blue.
/:cl: