Quality of Life, Spotters get to cloak and spot separately now (#9905)
# About the pull request

Changes the ability to activate spotting with the Spotter's Laser
Designator to Specialist Ability 2, removing the bloat of having the
ghillie suit cloak and the laser designator's ability bound to the same
button.

# Explain why it's good for the game

Quality of life addition for spotters to be able to cloak and spot
separately without having both abilities tied to the same button.

# Testing Photographs and Procedure
N/A

# Changelog



:cl:
qol: Spotter's spotting changed to Spec Ability 2 for smoother use of
both spotting and cloaking as a spotter.
/:cl: