Fax machine checks (#9967)
# About the pull request

This PR simply fixes a couple minor things with the fax machine:
- Should you manage to click fast enough or modify the UI, the backend
didn't check if the cooldown was finished before sending another fax
- Should you manage to get the ui to attempt to eject paper when there
is nothing inserted, it wasn't ending the call.

# Explain why it's good for the game

Some reason the double fax send was important enough for copilot to take
a stab at, so here's how to do it properly.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed the possibility of fax machines sending faxes on cooldown and
ejecting paper when there is no paper
/:cl: