stops dropship solid turfs being weedable (#9922)
# About the pull request

As title.

# Explain why it's good for the game

It is quite common for xeno weeds to spread under solid turfs on the
dropship which aren't accessible, and this can often lead to the queen
eye hiding in places marines can't remove, along with other niche
things.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: Dropship walls and other solid turfs cannot be weeded. This does
not affect dropship internals.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>