Dropship Lock now takes 5 seconds (#9578)
# About the pull request

As title

# Explain why it's good for the game

I dislike seeing queens running in on a screech and one-click locking
the dropship down then running off again. It is very frustrating to play
against and to observe, and doesn't make much sense to me, especially
with how long it takes to override it as a pilot.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Locking the dropship as Queen now takes 5 seconds.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>