Fixes a runtime namely with lessers who can die during Life (#9899)
# About the pull request

This PR simply fixes a runtime where because a lesser may die when
on_life() is called here:

https://github.com/cmss13-devs/cmss13/blob/ab882041cc096d231803990f2d57629f09758695/code/modules/mob/living/carbon/xenomorph/life.dm#L20-L34

they may get a null loc, and the existing code used that loc (but later
was checking the loc). So that checking is just moved up in the proc.

# Explain why it's good for the game

Fixes 

![image](https://github.com/user-attachments/assets/cad24595-07a7-4802-ab44-d79ee57c2eaf)

# Changelog
:cl: Drathek
fix: Fixed a runtime with lessers on death
/:cl: