Synth radio fix for SOF insert (#9916)
# About the pull request

new human new human new human.

Fixes some weird spawning with the Synth preset for the SOF insert on
Soro. The rest is probably unnecessary, but might as well.
# Explain why it's good for the game

Synth on the SOF insert for Soro was spawning with a radio that had the
default channel (Almayer) which was weird and unintended. Also took the
opportunity to adding (new_human) to all my presets from various PR's,
although probably completely unnecessary, ah well.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Zenith

fix: fixed some preset issues with radios spawning in nullspace on the
SOF insert for Soro.

/:cl: