attempt to fix the revolver's iff scope deleting ammo (#9943)
# About the pull request
see title, seems to work in testing. Confirmed russian roulette still
works as well. This is supposed to fix #8678 , I'm not sure how to link
all that stuff nicely.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
fixes good

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
fix: fixes the revolver iff scope bug
/:cl: