Destroying The Old Holdout Pistol (#9986)
# About the pull request

This PR destroys the old 'holdout pistol' with the weird blue finish and
ivory grip, and replaces it with two separate guns. The original holdout
pistol was clearly designed to be some sort of .22 plinker type gun,
think a ruger or some sort of target shooting gun. But it was small
enough to fit in a boot. I found that a bit odd, so I split it into two
guns that take different traits from the original gun.
(For reference, this is the original)
<img width="86" height="84" alt="image"
src="https://github.com/user-attachments/assets/09201a5a-e915-4e61-bbb2-262888251c3c"
/>


The main weapon, the W62 'Whisper', is a .22 plinker made by Spearhead
Armory. It is primarily used for shooting rats, target practice, and
some other things due to one specific trait it has. The Whisper, as its
namesake suggests, is integrally suppressed. It also has tritium
glow-in-the-dark sights, because they look cool. It is a small weapon,
and cannot fit in boots, but it will fit in your pocket. The 62 in the
designation is in reference to the year it was put on shelves, 2162. The
design is inspired by several different .22 plinkers, but it's primarily
based on the Colt Woodsman, with the slide only covering the back half
of the barrel and all such.
<img width="442" height="288" alt="image"
src="https://github.com/user-attachments/assets/f2acbd6b-9de3-4a73-bd0a-49137a2cbab7"
/>


It also has a custom variant that more faithfully represents the old
artstyle, 100% created by Zenith. I just posted a screenshot of the
thing, and he faithfully recreated it 1:1 and recolored it completely
unprompted. Absolute madlad.
<img width="431" height="264" alt="image"
src="https://github.com/user-attachments/assets/a06153dd-dbed-4c7a-a2ee-374b957d4b0f"
/>


The second weapon is the AC71 Action. It is a .380 ACP backup gun meant
to be used and then thrown away once expended. It is a tiny weapon, and
can fit in boots. For some reason, CMB deputies currently get the
holdout pistol, and I have no idea why. The Action replaces it in their
inventories. The 71 refers to the year it was put on shelves, 2171.
Design wise, Its based on a number of different holdout pistols found
online, but it most accurately resembles something like a mix between a
PPK and a P10 Grendel. Really it's its own design.
<img width="316" height="227" alt="image"
src="https://github.com/user-attachments/assets/5e71395a-25dd-4115-bdaf-866f48ac86fd"
/>

In most instances that the Holdout pistol was used, I replaced it with
the Whisper, but if it specifically needed to be a tiny item, I replaced
it with the Action. For things like crate loot pools, I simply added the
Action alongside the whisper to them. I didn't add the Action to any
maps, they will all now be the Whisper since it uses the same pathway as
the old Holdout pistol.

I also added both pistols to the loadout, since they're neat pistols and
it'd be good to see them used a bit more. It'd be great for RP and style
for some of the shipside roles, frankly.

I tested all of these changes on local, no mapping changes were done.
Inhands and Lineart for both guns have been added.



<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

This is good for the game because the old Holdout pistol didn't make
much sense, had an old sprite, and was generally an unfinished
afterthought. This fleshes it out and revitalizes it as two new guns.
And who doesn't want more cool guns?
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

<img width="335" height="147" alt="image"
src="https://github.com/user-attachments/assets/822391e2-4d5e-4f14-bf11-8b364372d4e1"
/>
<img width="321" height="137" alt="image"
src="https://github.com/user-attachments/assets/c45f59ea-ecdf-4a6a-bad4-ec151cb983a4"
/>
<img width="249" height="501" alt="image"
src="https://github.com/user-attachments/assets/63acc242-1621-4512-aea0-f41df6944583"
/>


</details>


# Changelog
:cl:
add: Added two new pistols, the Action and Whisper
del: Removed the old holdout pistol and replaced it with the Action and
Whisper
soundadd: Added a new firing sound for the Action pistol
imageadd: Added new sprites for the Whisper and Action pistols,
including Inhands
imageadd: Added a new sprite for the Action pistols magazines
imagedel: Deleted the old Holdout Pistol sprite
/:cl: