Minor Fix: Disposals no longer have no icon_state when loaded by nightmare (#10037)
# About the pull request
On nightmare loading, icon_state gets set to the base_icon_state without
any checking if it's null.
Slaps a check on it to prevent that. Invisible disposals pipes begone.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Uhhhhhh pipage?
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Nightmare disposal pipes no longer end up with null icon_states on
load.
/:cl: