Minor GAU fixes (#9938)
# About the pull request

this is a fix as it sets values to how I inteanded and how I said I am
changing them in the GAU rework. it makes spread of both ammo types 3
instead of 4 on AP ~~and ammo hits random bodypart to prevent broken
organs on every shot~~ (should already be the case)

# Explain why it's good for the game

makes GAU ammo work as inteanded ~~and spreads out its damage to otehr
bodyparts~~


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: reduces AP GAU ammo spread to 3 tile from 4 as intended
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>