Last xeno is unable to heal/rest, gets fullhealed and told to fight (#9936)
# About the pull request
See title, should hopefully curb roundend where it's a runner/lurker
running around and spending 45s healing between fights.

If I was willing to put in a bit more effort, the xeno could probably
have certain abilities (burrow, invis) disabled like marine specs,
alongside perhaps passive health loss, but I don't feel like it.

I first proposed this back in 2023 and most people were in agreement
with it, but at the time I wasn't really a contrib
https://forum.cm-ss13.com/t/final-xeno-gets-told-to-fight-less-stalling-more-fighting/4306


<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The last xeno should be fighting, not scampering around and healing for
10 minutes until an admin decides to abomb them. The faster they die the
faster the next round starts and the faster the 50+ players sitting in
dchat get to play again

The rejuv is just so that they can have one last hurrah and fight
marines from full health.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: The last living xeno is now fully healed, but is unable to
heal/rest at all. FIGHT AND DIE.
/:cl: