New Queen Sprite by Drawsstuff (#9419)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

Replaces the current Queen sprite with a custom made one. Credits for
the sprite goes to a list of devs and artists:

[CREDITS] 
medjic : For their work on a complete queen resprite for RUTGMC which
served as the visual basis and guideline for this resprite
Byte and mossymothling : For the first rough draft passes on the front
and side sprites and the visual guideline for the backsprite
Zenith00000 / Zenith / zenith_zero : For completely redoing the head on
the side sprite, as well as helping with the 2nd version's shading and
colors.
Buchse / Clockworkdoorhinge / Xenopera : For helping create the first
draft of the ovipositor sack
Blundir: For refactoring xeno wound overlays entirely, allowing the
larger canvas size to be used for wounds.
Drawsstuff : All final sprites including ovipositor and acid animations,
directionals, and all states such as wounded, weeded, stunned, dead, and
sleeping.

All I did was organize the reviving of the project as well as putting
together the pull request.

The sheer scale of her side profile means she does not fit as neatly on
a tile as before, but as both of her feet do land on the same tile, it
should be fine. Her size itself also makes it easier for marines to
click her sprite for weapon fire, as a tradeoff.

# Explain why it's good for the game

Queen's sprite is good, but doesn't represent just how unique a Queen is
compared to her hive. She is massive and stands out visually, and this
new sprite seeks to represent that in game.

Instead of just looking like a tall drone with a fancy crest, she truly
looks nothing like any other xeno, as she should.

# Testing Photographs and Procedure

Front profile

![image](https://github.com/user-attachments/assets/f0d9de0f-3dc5-4c56-b3fa-39cc489697df)
Ovi

![image](https://github.com/user-attachments/assets/f1f640be-066a-4219-bc88-edf944ac6299)
Ignore the fucky layering with the crusher and prae. Image meant to be a
side profile comparison.

![image](https://github.com/user-attachments/assets/66cf50da-7965-4be4-9d05-abc30de652a9)
Wound overlays

![image](https://github.com/user-attachments/assets/94728911-c269-4e4c-8426-8acfc93ec3f3)


</details>


# Changelog

:cl:
imageadd: added a new queen sprite
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->