The ID modification console doesn't allow editing by users other than the original (#9968)
# About the pull request

The ID modification console now checks that the person using it is the
same as the one who originally logged in.

# Explain why it's good for the game

This is only really used by people who want to use it to be shitters on
the ship and causes lots of headaches for both marine leadership and
staff.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: the ID modification console does not allow modification from other
users
/:cl:

---------

Co-authored-by: khansworth esq. <fourkhan@da.khan>