Minor Fix: No longer Click drag items whilst Hauled, Linter fix for SHOULD_CALL_PARENT in attack_self (#9762)
# About the pull request
Adds a check for hauling when you click drag an item to hand;
~~Currently just returns doesn't check type.~~ Edit: Does check for gun
or weapon. Doesn't work for click dragging weapon/gun to hand from non
storage, I.E. M41 in back slot. Will work for pistols in a holster or
other storage though.
Sprinkles more checks around for Alt+Clicking storage, internal
storages, and defaults to opening the storage if they are hauled but
don't have the allow inv haul flag. Instead of doing nothing.
Moves a check below a Linter needed thing so it doesn't break the Linter
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
We need Linters for linting
Probably shouldn't let items be grabbed that you can't get by clicking
on them.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas, Ediblebomb, Drathek
fix: Adds a haul check for click dragging, moves a Lint Proc to the top
for Linting; Fixes the things that would get caught in the fixed lint.
Left clicking storages whilst hauled will open them instead of doing
nothing.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>