Scout now has a built in radio telephone in its helmet (#9893)
# About the pull request
Part of the expansion i promised post scout weapon nerf

# Explain why it's good for the game

Scout is a very essential role to the game when picked, its very strong
as a spotter due to its faster binoc acquistion time for both cas and
coordinates for mortar.

A lot of its lost when comms are lost, while yes that ssupposed to be a
big drawback i believe while trying to move scout from giga fragger to
more of a support role, this will help in the transition.

Also i think it will be awesome to call scout players when they are in
the backline and get them killed

I've changed this into the helmet instead of the backpack at the request
of Blundir, the only difference from RTO backpacks is you cant call and
give it to somebody, so you have to give them the helmet itself.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Scout helmet now has the same functions as an RTO pack,
allowing you to call or be called. However, you cannot hand the
telephone to anyone and you must give them the helmet if you want them
to use it.
add: Adds radio backpack equivelant for helmets, nothing uses these
right now other then scout.
imagedel: grenadier helmet no longer has an antenna, scout helmet does
soundadd: New sounds for the scout helmet-phone
imageadd: New sprites for the scout helmet and microphone
/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>
Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>