Multi Z Almayer [MDB IGNORE] (#9590)
# About the pull request

Makes the USS Almayer multiz, to do this I had to rearrange a lot of
stuff to get everything where it needs to be, I tested it a little prob
gonna need a TM

Also I changed the BBall court bc I felt like it and it looks grimmer

This PR also adds 'The underdeck' an inaccessible part of the Almayer
containing, Lower Cargo, Vehicle bay and hangar. Allowing ghosts and
people watching security cameras a little extra immersion and placing
the groundwork for dropship airlocks should that be ported since it
would need a Zlevel below the Almayer anyway. To that end this PR
removes the vehicle and cargo elevators from the Admin level, placing
them in their correct position under the almayers lower deck

# Explain why it's good for the game

Making use of multiz is cool instead of using Fake Z and teleports and
projectors

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:SpartanBobby
maptweak: Makes USS Almayer Multiz
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>
Co-authored-by: forest2001 <forest2001@cm-ss13.com>