New firearms max level, change several firearms skills. (#10271)
# About the pull request

This PR introduces a new firearms skill named "Skilled" and change
"Expert" to "Master", the Master skills is the old Expert and the
Skilled works as midle ground between Trained and Master, while Trained
multiplier is 1 and Master multiplier is 2(Exactly like the Trained and
Expert worked), Skilled has a multiplier of 1.5, giving another set of
possibilities to the contributors, since the difference from Trained to
Expert was massive, not everyone wanted to give their new survivor,job
or ERT the skill level of Expert, so there is a new workaround called
Skilled, this PR also changes alot of the skills level through the game,
Master being reserved to deathsquads, special individuals or very high
trained specialists, while Skilled is more given through since it is not
as strong.

This is both a buff and a nerf, Survs that had the previous Master were
nerfed to Skilled and there is not a single surv with the Master skill,
now Skilled was given more out to the survs specially some like IASF or
FORECON, since they got a level of training justifying that, PMCs survs
and Goons survs that had the level of Master were reduced to Skilled.

Marines CO,XO,MST,SEA,SPEC,SL recieved changes as well, mainly making
CO,XO,SPEC,SL Firearm skills now set to Skilled, SEA from the previous
Expert to Skilled and MST continue with their max level of firearms.

The entire UPP with the exception of Kommando and the Specialist were
given the level of Skilled from the previous Expert level, CMB Marshal
level was reduced to Skilled while the other CMB were increased to
skilled including the RCU members, CLF all continue with the same level
of training with the exception being CLF Cell Commander from Expert to
Master and the CLF Specialist with their level of firearm being Skilled

The Majority that had the level of Expert is now Skilled and only a few
got the level of Master firearms.


# Explain why it's good for the game

I always felt that the "3" levels aka Civilian 0, Trained 1, Expert 2,
were not good enough for the amount of factions and different groups the
game has, since the bridge between Trained and Expert was so high, jobs
that didn't had the Expert level were given solely because trained was
too weak and while other jobs that definetly had a skill level higher
than trained were stuck with that since Expert was too strong to them.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: New level of firearms skill
balance: Change the firearms skill of several jobs
/:cl: