Adds an antiZed autoinjector (#10341)
# About the pull request

Adds a single-use autoinjector that is pre-filled with five units of
antiZed. This will cure someone of the virus if used on themselves or
someone else. The injector is skillless.

# Explain why it's good for the game

Gives more options for event runners to distribute the cure in a more
controlled manner, as well as allowing non-medical players to use it.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

<img width="216" height="202" alt="Capture"
src="https://github.com/user-attachments/assets/9341b033-7d25-4db8-9519-bd29aad92c80"
/>

</details>


# Changelog
:cl:
add: An antiZed autoinjector has been added to the game, a single-use
device that contains a cure for the zombie pathogen.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>