CO surv bugfixes (#10270)
# About the pull request
Two things:
sets CO equipment for pmc insert spawners to their own default equipment
to prevent CO survs from spawning in the PMC inserts as CMB. If and when
a PMC surv CO is implemented, this can be set to an actual PMC CO surv,
in the meantime, the CMB marshal shouldn't be materializing in the PMC
insert with flavortext about being a CL or whatever.

also, allows CLF co survs to actually spawn, as the default slots for CO
surv is 0 unless you have an equipment preset for them in the map json,
which most maps do not.

# Explain why it's good for the game


bugs bad

# Testing Photographs and Procedure

too hard

# Changelog
:cl:
fix: CLF CO survs now actually spawn, CO survs in PMC insert will no
longer be CMB
/:cl: