UACQS Presets (#10348)
# About the pull request

Adds quick presets for the UACQS (United Americas Commission for Quality
and Standards) which is definitely not a cover for the CIA.

# Explain why it's good for the game
Adds a little bit of simplicity to a semi-frequent RP element I do.

# Changelog
:cl:
add: Added equipment presets for UACQS.
add: Added CIA roles to the crew monitor.
add: Added a CIA clothing vendor.
add: Added lockable versions of the blue and black leather satchels.
add: Added a variable that determines if signs can be deconstructed, and
forbids it on no-smoking signs and the RO plaque.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>