Seethrough for queen (#10259)
# About the pull request

Makes it so you can press an ability to make yourself see through as the
queen, since the recent sprite is much bigger facing some areas.

port of  https://github.com/tgstation/tgstation/pull/77361

This is currently only on queen, but you can assign the ability and
component to anyone and it will work with any xeno

# Explain why it's good for the game

Sprites shouldnt be stopping people from playing the game


https://github.com/user-attachments/assets/71a5729d-07dc-42fd-9a91-28df1da2cd25



# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: Added a transparent sprite ability for the queen
/:cl: