Fixes the Almayer wall blended shutters not working sideways. (#10412)
# About the pull request

As title, this fixes an issue with missing sprites that allows the wall
blended shutters to be used in east-west facing positions

# Explain why it's good for the game

Makes it functional when it is supposed to be.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Blended wall shutters can now be used properly in east-west
directions.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>