Prevents climbing with C4 and subtypes in-hand (#10452)
# About the pull request

This means you don't try to climb + plant C4 at the same time. 

You could accomplish this before by being on harm-intent, but if you're
holding C4 you're likely trying to plant it, and not climb.

Fixes #10415

# Explain why it's good for the game

Maybe it isn't, let me know and we can close this or find a different
solution. (The easiest of course being 'just be on harm-intent', but :/)
The code for this isn't very elegant either. Climbing happens before
C4's afterattack, and I don't see a good way to cancel an in-progress
climb via the C4 afterattack. (Not sure we'd even want that to be fair)

# Testing Photographs and Procedure
Tested by trying to climb a wall with C4 in my hands.
Didn't climb, could plant C4


# Changelog

:cl:
qol: You can no longer climb with C4 in your hands, as you're more
likely to want to plant it on the wall.
/:cl: