Sentries now correctly process mobs inside inventories (#10458)
Update sentry.dm
# About the pull request

If you put a CLF member inside a closet and put them infront of the
sentry, the sentry will shoot the closet as intended.

But if you then move the closet behind the (non-omnidirectional) sentry,
it will keep firing.
The math assumes that the possible target has x and y coordinates.
Replaces all references to .x and .y with the .x and .y of the turf
returned from calling get_turf() with the target.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Sentries will correctly handle mobs in inventories.
/:cl: