forecon crew monitor (#10478)
# About the pull request

Puts forecon on the crew monitor.

# Explain why it's good for the game

They're on the monitor by default due to faction settings, but they all
appear in a jumble at the bottom.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added designated Forecon listing on the crew monitor.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>