Fixes inhand sprites for PKD M2049 (#10385)
It's a unique little revolver and you can't replace the ammo for it.
It has a tiny damage increase vs the base M44, but otherwise stats are
identical. I don't think it poses any balance risks.

# Explain why it's good for the game

Bugs bad

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixes the inhand sprites for all three PKD revolvers to use their
existing placeholders.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>