OT Lab changes (#10480)
# About the pull request
This Pull request is to address some of the key issues that I have found
with the current layout of OT during the time I have been playing it.
This is also a start to the overhaul of OT.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Now, for why it is good for the game currently, the OT storage is very
cramped, and when you have more than one person playing OT, you get in
each other's way, especially with the layout of the chem tanks. With
this change, the left-hand side is completely equipped with chemicals to
work on paraform and formaldehyde. On the right-hand side, it has the
chems to make incendiary nades such as napalm. While also on both sides
staging them in the way of progression, making the production of nades
more streamlined, logical, and easier to learn for new Players.
# Testing Photographs and Procedure
<img width="546" height="451" alt="Screenshot 2025-08-23 113220"
src="https://github.com/user-attachments/assets/81419f35-87a7-44fc-a3c2-4de643b66179"
/>
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins, you should add a changelog. If your
change does NOT meet this description, remove this section. Be sure to
properly label your changes in the changelog. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->
:cl:
maptweak: Tweaked OT lab to have a better layout for easier use
/:cl: