Fixes the eternal hangup loop (#10528)
# About the pull request

Fixes #8767 

Moves hangup_loop.stop() out of the mob loc check, to enable the loop to
stop even if the phone is thrown down the ASRS pit.

I haven't been able to find any instances of this needing to be inside
the check.

# Explain why it's good for the game

Throwing the phone into the ASRS pit (or otherwise destroying it without
it being on a person) probably shouldn't cause it to ring forever.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/9g-m9EHrRQc

</details>


# Changelog
:cl:
fix: Throwing the phone down the ASRS pit after being hung up on won't
cause it to loop forever.
/:cl: